my.vmware.com API client -- because why not?
This is the start of client lib for my.vmware.com.
It is pretty much a re-implementation of
- https://github.com/apnex/vmw-sdk/
- https://github.com/apnex/vmw-cli/

Logging in seems to work, some methods have been implemented.

I have no idea about the exact repsonse types, I have no idea where I
could find the API docs.

There are not many tests, this is pretty much a spike as of now.

If you want to run the e2e tests, make sure to set VMW_USER & VMW_PASS.